requirements.txt
aniso8601==8.0.0
boto3==1.10.2
botocore==1.13.2
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-RESTful==0.3.7
google-ads==3.3.0
google-api-core==1.14.3
google-auth==1.6.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
oauthlib==3.1.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
